+++ src/Blog/WebBundle/Controller/PostsController.php
@@ -54,7 +54,7 @@ class PostsController extends Controller {
 	 * @extra:Routes({
 	 *	@extra:Route("/post/{pid}/", requirements={"pid" = "\d+"}, name="_posts_show"),
 	 *	@extra:Route("/post/{pid}/#comments", name="_posts_show_comments"),
-	 *	@extra:Route("/post/{pid}/edit-comment/{cid}", requirements={"pid" = "\d+", "cid" = "\d+"}, name="_posts_show_comments_edit")
+	 *	@extra:Route("/post/{pid}/edit-comment/{cid}", name="_posts_show_comments_edit", requirements={"pid" = "\d+", "cid" = "\d+"})
 	 * })
 	 * @extra:Template()
 	 */
@@ -89,21 +89,25 @@ class PostsController extends Controller {
 			return $this->redirectGenerate('_posts_show', array('pid' => $post->getId()));
 		}
 
-		$this->addTitle('Posts', 'Add post');
+		$this->addTitle('Posts', 'Add');
 		return array('form' => $form);
 	}
 	
 	/**
-	 * @extra:Route("/post/{pid}/edit", requirements={"pid" = "\d+"}, name="_posts_edit")
+	 * @extra:Route("/post/{pid}/edit", name="_posts_edit", requirements={"pid" = "\d+"})
 	 * @extra:Template()
 	 */
 	public function editAction($pid) {
 		$em = $this->getEm();
 		$post = $em->find('Blog\\WebBundle\\Entity\\Posts', $pid);
-		// not found
+		
 		if (!$post) {
 			throw ExceptionController::notFound('The post does not exist.');
 		}
+		if (!$post->canEdit($this->getUser())) {
+			throw ExceptionController::forbiden();
+		}
+		
 		$form = PostsAddForm::create($this->get('form.context'), 'posts_edit');
 		
 		$form->bind($this->get('request'), $post);
@@ -115,23 +119,57 @@ class PostsController extends Controller {
 			return $this->redirectGenerate('_posts_show', array('pid' => $pid));
 		}
 
-		$this->addTitle('Posts', 'Edit post', $post);
+		$this->addTitle('Posts', 'Edit', $post);
 		return array('form' => $form);
 	}
 	
 	/**
-	 * @extra:Route("/post/{pid}/delete", name="_posts_delete")
+	 * @extra:Route("/post/{pid}/delete", name="_posts_delete", requirements={"pid" = "\d+"})
 	 */
 	public function deleteAction($pid) {
 		$em = $this->getEm();
 		$post = $em->find('Blog\\WebBundle\\Entity\\Posts', $pid);
-		// not found
+		
 		if (!$post) {
 			throw ExceptionController::notFound('The post does not exist.');
 		}
+		if (!$post->canDelete($this->getUser())) {
+			throw ExceptionController::forbiden();
+		}
+		
 		$em->remove($post);
 		$em->flush();
 		
 		return $this->redirectGenerate('_posts');
+	}
+	
+      /**
+	 * @extra:Route("/posts/user/{uid}/{pageLabel}/{page}/", name="_posts_users", requirements={"page" = "\d+", "uid" = "\d+", "pageLabel" = "page"}, defaults={"page" = 1, "pageLabel" = "page"})
+	 * @extra:Template()
+	 * 
+	 * @todo SQL_CALC_FOUND_ROWS
+	 */
+	public function userAction($uid, $pageLabel, $page) {
+		$em = $this->getEm();
+		$user = $em->find('Blog\\WebBundle\\Entity\\Users', $uid);
+		// not found
+		if (!$user) {
+			throw ExceptionController::notFound('The user does not exist.');
+		}
+		$this->addTitle('Posts', 'User', $user);
+		
+		$qb = $em->createQueryBuilder()
+			   ->select('p')
+			   ->from('Blog\\WebBundle\\Entity\\Posts', 'p')
+			   ->where('p.uid = :uid');
+		$q = $qb->setParameters(array('uid' => $user->getId()))->getQuery();
+		
+		// paginator
+		$paginatorAdapter = $this->getPaginatorAdapter()
+						 ->setQuery($q)
+						 ->setDistinct(true);
+		$paginator = $this->createPaginator($paginatorAdapter);
+
+		return array('paginator' => $paginator, 'user' => $user);
 	}
 }
